+++ app/Message.php
@@ -66,7 +66,7 @@ class Message extends Model
         static::saved(function (Message $message) {
             if ($message->messageFiles != null && $message->messageFiles->isNotEmpty()) {
                 $mid = Message::where('id', $message->id)
-                    ->where('user_id', me()->id)
+                    ->where('user_id', $message->user_id)
                     ->where('jidfrom', $message->jidfrom)
                     ->first()
                     ->mid;
@@ -123,7 +123,6 @@ class Message extends Model
                     ->where('user_id', me()->id)
                     ->where('jidto', $jid)
             );
-
         return $query->select('*')->from(
             $jidFromToMessages,
             'messages'
@@ -185,7 +184,7 @@ class Message extends Model
 
     public static function findByStanza(?\SimpleXMLElement $stanza = null, ?\SimpleXMLElement $parent = null): Message
     {
-        $jidfrom = baseJid((string)$stanza->attributes()->from);
+        $jidfrom = bareJid((string)$stanza->attributes()->from);
 
         if (
             $stanza->attributes()->xmlns
@@ -194,7 +193,7 @@ class Message extends Model
             return self::firstOrNew([
                 'user_id' => me()->id,
                 'stanzaid' => (string)$stanza->attributes()->id,
-                'jidfrom' => baseJid((string)$stanza->forwarded->message->attributes()->from)
+                'jidfrom' => bareJid((string)$stanza->forwarded->message->attributes()->from)
             ]);
         } elseif (
             $stanza->{'stanza-id'} && $stanza->{'stanza-id'}->attributes()->id
@@ -634,7 +633,7 @@ class Message extends Model
             if (isset($stanza->x->invite)) {
                 $this->type = 'invitation';
                 $this->subject = $this->jidfrom;
-                $this->jidfrom = baseJid((string)$stanza->x->invite->attributes()->from);
+                $this->jidfrom = bareJid((string)$stanza->x->invite->attributes()->from);
             }
         } elseif (
             isset($stanza->x)