+++ app/Widgets/Search/Search.php
@@ -84,8 +84,7 @@ class Search extends Base
                     $filters = DB::table('posts')->where('id', -1);
 
                     $filters = \App\Post::withMineScope($filters);
-                    $filters = \App\Post::withContactsFollowScope($filters);
-                    $filters = \App\Post::withCommunitiesFollowScope($filters);
+                    $filters = \App\Post::withFollowScope($filters);
 
                     $query->select('id')->from(
                         $filters,
@@ -99,7 +98,7 @@ class Search extends Base
                 $view->assign('posts', $posts);
             }
 
-            $contacts = Contact::suggest($key)->limit(10)->get();
+            $contacts = Contact::suggest($this->me)->limit(10)->get();
 
             if (validateJid($key) && $key != $this->me->id) {
                 $contact = new Contact;
@@ -137,7 +136,7 @@ class Search extends Base
     {
         $view = $this->tpl();
 
-        $users = Contact::suggest()
+        $users = Contact::suggest($this->me)
             ->limit(12)
             ->get();